move PriceGenerator to simulation/types to keep logic clean
